Extract advisor brief source fact helpers
Move advisor-brief source formatting, contributor ranking, and AI fact-bundle shaping out of advisor_brief_source.py while preserving the existing fact-bundle import path.

Evidence:

- python -m ruff check src/app/services/advisor_brief_source.py src/app/services/advisor_brief_source_fact_bundle.py src/app/services/advisor_brief_source_contributors.py src/app/services/advisor_brief_source_formatting.py tests/unit/test_advisor_brief_source.py tests/unit/test_refactor_quality_thresholds.py scripts/check_refactor_quality_thresholds.py: passed

- python -m ruff format --check src/app/services/advisor_brief_source.py src/app/services/advisor_brief_source_fact_bundle.py src/app/services/advisor_brief_source_contributors.py src/app/services/advisor_brief_source_formatting.py tests/unit/test_advisor_brief_source.py tests/unit/test_refactor_quality_thresholds.py scripts/check_refactor_quality_thresholds.py: passed

- python -m pytest tests/unit/test_advisor_brief_source.py tests/unit/test_advisor_brief_narrative.py tests/unit/test_refactor_quality_thresholds.py -q: 12 passed

- python -m mypy src/app/services/advisor_brief_source.py src/app/services/advisor_brief_source_fact_bundle.py src/app/services/advisor_brief_source_contributors.py src/app/services/advisor_brief_source_formatting.py: passed
